Re-implement linear fill
