+++ server/src/main/java/org/apache/iotdb/db/query/executor/FillQueryExecutor.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.query.executor.fill;
+package org.apache.iotdb.db.query.executor;
 
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.exception.StorageEngineException;
